[Tools] Set "target_os" in .gclient-xwalk based on .gclient.
Instead of requiring the XWALK_OS_ANDROID to be set, use the value that
is set in .gclient.

This should bring Crosswalk's behaviour more in line with Chromium's,
with all user-configurable gclient settings being fetched from .gclient
itself instead of depending on other sources such as environment
variables.

.gclient-xwalk now behaves as closely to .gclient as possible: the
solutions in each file are different, but other settings are the same.

XWALK_OS_ANDROID is still used if set for compatibility reasons, but
support for it will be removed in Crosswalk 8.