M53: Remove the use of ContentViewDownloadDelegate and DownloadInfo for now
As the ContentViewDownloadDelegate has been removed from content upstream
to let the embedder handle downloads, we need to do the same for crosswalk.
DownloadInfo has been moved from content/ to chrome/ as well.
For now just getting rid of the depndency but we may need to reimplement
the way we handle downloads.

adapt to,
https://codereview.chromium.org/2014803002